+++ tests/integration/models/WaterEnergy.py
@@ -32,9 +32,11 @@ from gamspy import Equation
 from gamspy import Model
 from gamspy import Parameter
 from gamspy import Set
+from gamspy import SolveStatus
 from gamspy import Sum
 from gamspy import Variable
 from gamspy.math import sqr
+from gamspy.exceptions import GamspyException
 
 
 def reformat_df(dataframe):
@@ -264,7 +266,13 @@ def main():
         sense="min",
         objective=TC + CC + WaterCost,
     )
-    DEDcostbased.solve()
+    try:
+        DEDcostbased.solve()
+    except GamspyException:
+        if DEDcostbased.solve_status == SolveStatus.TerminatedBySolver:
+            pass
+        else:
+            raise
 
     print(
         "Objective Function Value:  ", round(DEDcostbased.objective_value, 4)